requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
click==8.1.7
pytest==7.4.3
pytest-mock==3.12.0